Add files via upload
One of my two 'Platformer' projects, named the same as I did not realize this one was called Platformer, which is apparent by the 'Alchemy' name strewn around the project. There are many standard assets missing as there were too many to upload, but these have the main working code.